新形式のNPCバトルステートを追加した
