+++ test/js/game/is-stage-clear.test.js
@@ -1,6 +1,6 @@
 // @flow
-import {isStageClear} from "../../../../src/js/game/in-progress/npc-battle";
-import {EMPTY_PLAYER} from "../../../data/player";
+import {EMPTY_PLAYER} from "../../data/player";
+import {isStageClear} from "../../../src/js/game/npc-battle";
 
 const player = {...EMPTY_PLAYER, playerId: 'test-player'};
